requirements.txt
asgiref==3.11.1
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==26.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.6.0
